refactor(stk): change stk_init() return type to uint8_t
- Change function signature from int to uint8_t for clarity
- Makes return type intent explicit as status flag

+++ include/stk.h
@@ -2,6 +2,7 @@
 #define STK_H
 
 #include "stk_version.h"
+#include <stdint.h>
 #include <stdlib.h>
 
 /* Buffers */
@@ -54,7 +55,7 @@ typedef enum {
 	STK_MOD_RELOAD
 } stk_module_event_t;
 
-int stk_init(void);
+uint8_t stk_init(void);
 void stk_shutdown(void);
 size_t stk_module_count(void);
 size_t stk_poll(void);
